% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\docType{methods}
\name{asBED,GInteractions-method}
\alias{asBED,GInteractions-method}
\title{Coerce to BED structure}
\usage{
\S4method{asBED}{GInteractions}(x, keep.mcols = TRUE)
}
\arguments{
\item{x}{Generally, a tabular object to structure as BED}

\item{...}{Arguments to pass to methods

     The exact behavior depends on the class of ‘object’.

     ‘GRangesList’ This treats ‘object’ as if it were a list of
          transcripts, i.e., each element contains the exons of a
          transcript. The ‘blockStarts’ and ‘blockSizes’ columns are
          derived from the ranges in each element. Also, add ‘name’
          column from ‘names(object)’.}
}
\value{
A ‘GRanges’, with the columns ‘name’, ‘blockStarts’ and
        ‘blockSizes’ added.
}
\description{
Coerce the structure of an object to one following BED-like
conventions, i.e., with columns for blocks and thick regions.
}
\examples{
data(hic_example_data)
asBED(hic_example_data)
}

