% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{countsBetweenAnchors}
\alias{countsBetweenAnchors}
\alias{countsBetweenAnchors,GenomicInteractions,GRanges-method}
\title{Summarise Interactions between defined anchors}
\usage{
countsBetweenAnchors(x, y)

\S4method{countsBetweenAnchors}{GenomicInteractions,GRanges}(x, y)
}
\arguments{
\item{x}{A GenomicInteractions object}

\item{y}{A GenomicRanges object}
}
\value{
A GenomicInteractions object with annotated counts between anchors
}
\description{
Calculate the number of of paired-end reads mapping between a defined set of anchors.
This function will ignore counts present in the input data.
}

