% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/InteractionTrack-class.R
\docType{methods}
\name{drawGD,InteractionTrack-method}
\alias{drawGD,InteractionTrack-method}
\title{draws InteractionTrack}
\usage{
\S4method{drawGD}{InteractionTrack}(GdObject, minBase, maxBase,
  prepare = FALSE, subset = TRUE, ...)
}
\arguments{
\item{GdObject}{}

\item{minBase}{}

\item{maxBase}{}

\item{prepare}{}

\item{subset}{}
}
\value{
an InteractionTrack object
}
\description{
draws InteractionTrack
}

