% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{c}
\alias{c}
\alias{c,GenomicInteractions}
\title{Combine GenomicInteractions Methods}
\usage{
\S4method{c}{GenomicInteractions}(x, ..., ignore.mcols = FALSE,
  recursive = FALSE)
}
\arguments{
\item{x,}{... GenomicInteractions objects to be concatenated}

\item{ignore.mcols}{Logical, default FALSE, remove mcols in combined object.}

\item{recursive}{Not supported}
}
\description{
This method will fail if
}

