% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[}
\alias{[}
\alias{[,GenomicInteractions-method}
\title{Standard subsetting methods for GenomicInteractions objects}
\usage{
\S4method{[}{GenomicInteractions,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{A genomicInteractions object}

\item{i}{A numeric, logical or Rle vector}
}
\value{
A GenomicInteractions object containing only the features specified by `i`.
}
\description{
Standard subsetting methods for GenomicInteractions objects
}

