% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicInteractions-class.R
\name{GenomicInteractions}
\alias{GenomicInteractions}
\title{Function to create a GenomicInteractions object}
\usage{
GenomicInteractions(anchor1, anchor2, counts, ...)
}
\arguments{
\item{counts}{An integer vector, defaults to 1.}

\item{...}{Additional data to be added to mcols}

\item{anchor_one, anchor_two}{GRanges objects.}

\item{experiment_name}{Experiment name.}

\item{description}{Description of experiment.}
}
\value{
a GenomicInteractions object
}
\description{
Create GenomicInteractions objects from two GRanges ojects.
}
\examples{

library(GenomicRanges)

anchor.one = GRanges(c("chr1", "chr1", "chr1", "chr1"), IRanges(c(10, 20, 30, 20), width=5))
anchor.two = GRanges(c("chr1", "chr1", "chr1", "chr2"), IRanges(c(100, 200, 300, 50), width=5))
interaction_counts = sample(1:10, 4)
test <- GenomicInteractions(anchor.one, anchor.two, counts=interaction_counts)

}

