% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_binom_ligation_threshold}
\alias{get_binom_ligation_threshold}
\title{get self ligation threshold with binomial test}
\usage{
get_binom_ligation_threshold(GIObject, max.distance = 20000, bin.size = 500,
  p.cutoff = 0.05, adjust = "fdr", plot = TRUE)
}
\description{
get self ligation threshold with binomial test
}

