% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_self_ligation_threshold}
\alias{get_self_ligation_threshold}
\title{Get self ligation threshold with SD method from Heidari et al}
\usage{
get_self_ligation_threshold(GIObject, bins = 100, distance_th = 4e+05,
  plot = TRUE)
}
\description{
Get self ligation threshold with SD method from Heidari et al
}

