% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/InteractionTrack-class.R
\docType{class}
\name{InteractionTrack-class}
\alias{InteractionTrack-class}
\title{A class to hold chromatin interaction data for a specific genomic region.}
\description{
A class to hold chromatin interaction data for a specific genomic region.
}
\section{Slots}{

\describe{
\item{\code{plottingFunction}}{function}

\item{\code{variables}}{list}

\item{\code{chromosome}}{}

\item{\code{stacking}}{character

InteractionTrack is a specific Gviz-derived class for enabling the visualisation of chromatin interaction data.
The InteractionTrack class allows interactions on a specified chromosome to be visualised by examining interactions
between anchors as bezier curves.

Several additional display parameters (i.e. \rcode{displayPars(foo)=list(...) }are defined for this class, including
\code{plot.anchors} which can be used to specify whether anchors are to be drawn. \code{col.anchors} which can be used
to alter the colour of these anchor elements. The value of \code{plot.outside} determines whether or not interactions
which span outside of the window are to be plotted, and \code{col.outside} defines the colour of these interactions.
Similarly \code{plot.trans} determines whether trans-interactions are plotted and \code{col.trans} specifies the colour
of trans-interactions. By default, the line width of an arc representing an interaction is proportional to the number
of reads/counts supporting that interaction. Instead of using the counts to define this, the line width can be set to
be proportion to either \code{fdr} or \code{p.value} using the \code{lwd.interactions} display parameter.}
}}

