% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{sort,GenomicInteractions-method}
\alias{sort,GenomicInteractions-method}
\title{Sort GenomicInteractions Object}
\usage{
\S4method{sort}{GenomicInteractions}(x, decreasing = FALSE)
}
\arguments{
\item{x}{GenomicInteractions Object}

\item{decreasing}{A logical indicating sort order}
}
\value{
A sorted GenomicInteractions object
}
\description{
This method will sort a GenomicInteractions object by first
arranging all interactions start on the lower-ordered anchor;
for trans-chromosomal interactions this is the anchor on the
lower ordered chromomsome (defined by the seqlevels factor);
and then by the position of the first anchor.
}

