% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicInteractions-class.R
\docType{methods}
\name{updateObject,GenomicInteractions-method}
\alias{updateObject,GenomicInteractions-method}
\title{updateObject method for GenomicInteractions 1.3.7 and earlier}
\usage{
\S4method{updateObject}{GenomicInteractions}(object, ..., verbose = FALSE)
}
\description{
updateObject method for GenomicInteractions 1.3.7 and earlier
}

